Make poppler_movie_get_volume() return double 0.0-1.0

+++ glib/poppler-movie.h
@@ -78,7 +78,7 @@ PopplerMoviePlayMode poppler_movie_get_play_mode (PopplerMovie *poppler_movie);
 POPPLER_PUBLIC
 gboolean             poppler_movie_is_synchronous (PopplerMovie *poppler_movie);
 POPPLER_PUBLIC
-gint                 poppler_movie_get_volume (PopplerMovie *poppler_movie);
+gdouble              poppler_movie_get_volume (PopplerMovie *poppler_movie);
 POPPLER_PUBLIC
 gdouble              poppler_movie_get_rate (PopplerMovie *poppler_movie);
 POPPLER_PUBLIC
